@@ -1,14 +1,22 @@
 {
   "name": "nfl-telegram-bot",
-  "version": "2.1.4",
+  "version": "3.0.0",
   "description": "A Telegram bot that provides the latest news about the NFL",
   "repository": "github:jpmoura/nfl-news-for-telegram",
-  "main": "index.ts",
+  "main": "src/index.ts",
+  "packageManager": "yarn@1.22.17",
+  "engines": {
+    "node": ">= 12.0.0"
+  },
   "scripts": {
+    "prestart": "yarn build",
     "start": "node ./dist/index.js",
-    "tsc": "tsc",
-    "commit": "git-cz",
-    "sonar":  "node sonar-project.js"
+    "prebuild": "npx rimraf dist",
+    "build": "npx tsc --project ./tsconfig.build.json",
+    "test": "jest",
+    "test:cov": "jest --coverage",
+    "test:ci": "yarn test:cov --maxWorkers=2",
+    "lint": "npx eslint --ext .ts src/ tests/"
   },
   "keywords": [
     "NFL",
@@ -33,28 +41,37 @@
     "axios": "^0.21.2",
     "cheerio": "^1.0.0-rc.5",
     "dotenv": "^6.1.0",
+    "inversify": "^6.0.1",
     "nedb": "^1.8.0",
     "node-schedule": "^1.3.0",
+    "reflect-metadata": "^0.1.13",
     "telegraf": "^3.25.0",
+    "tslog": "^3.3.1",
     "twit": "^2.2.11"
   },
   "devDependencies": {
-    "@types/cheerio": "^0.22.22",
     "@types/dotenv": "^8.2.0",
+    "@types/faker": "^5.5.9",
+    "@types/inversify": "^2.0.33",
+    "@types/jest": "^27.0.3",
     "@types/nedb": "^1.8.11",
     "@types/node-schedule": "^1.3.1",
     "@types/twit": "^2.2.28",
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
     "commitizen": "^4.2.2",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^7.14.0",
-    "eslint-config-airbnb-typescript": "^12.0.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.3",
+    "eslint": "^8.6.0",
+    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^25.3.4",
+    "faker": "^5.5.3",
     "husky": "^4.3.4",
-    "jest": "^26.6.3",
-    "sonarqube-scanner": "^2.8.0",
-    "typescript": "^4.1.2"
+    "jest": "^27.4.7",
+    "jest-mock-extended": "^2.0.4",
+    "jest-sonar": "^0.2.12",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^27.1.2",
+    "typescript": "^4.5.4"
   }
 }
